restructure: move jsonLd generation from app/ to seo/
