+++ app/[slug]/page.test.tsx
@@ -124,100 +124,6 @@ describe('DynamicRoute page', () => {
     vi.clearAllMocks()
   })
 
-  describe('JSON-LD structured data', () => {
-    it('renders JSON-LD script with Article schema', async () => {
-      const mockPost = {
-        id: '123',
-        slug: 'test-post',
-        title: 'Test Post',
-        description: 'Test description',
-        firstPublished: '2024-01-15',
-        lastEditedTime: '2024-01-20T10:00:00.000Z',
-        featuredImage: 'https://example.com/image.jpg',
-        blocks: [],
-        prevPost: null,
-        nextPost: null,
-      }
-
-      vi.mocked(getPost).mockResolvedValue(Ok(mockPost))
-
-      const params = Promise.resolve({ slug: 'test-post' })
-      const jsx = await DynamicRoute({ params })
-      render(jsx)
-
-      const scriptTag = document.querySelector('script[type="application/ld+json"]')
-      expect(scriptTag).toBeInTheDocument()
-
-      const jsonLd = JSON.parse(scriptTag?.textContent ?? '{}')
-      expect(jsonLd).toEqual({
-        '@context': 'https://schema.org',
-        '@type': 'Article',
-        headline: 'Test Post',
-        description: 'Test description',
-        datePublished: '2024-01-15',
-        dateModified: '2024-01-20T10:00:00.000Z',
-        author: {
-          '@type': 'Person',
-          name: 'Michael Uloth',
-        },
-        image: 'https://example.com/image.jpg',
-        url: 'https://michaeluloth.com/test-post/',
-      })
-    })
-
-    it('uses default OG image when post has no featured image', async () => {
-      const mockPost = {
-        id: '123',
-        slug: 'no-image',
-        title: 'Post Without Image',
-        description: 'No image',
-        firstPublished: '2024-01-15',
-        lastEditedTime: '2024-01-15T00:00:00.000Z',
-        featuredImage: null,
-        blocks: [],
-        prevPost: null,
-        nextPost: null,
-      }
-
-      vi.mocked(getPost).mockResolvedValue(Ok(mockPost))
-
-      const params = Promise.resolve({ slug: 'no-image' })
-      const jsx = await DynamicRoute({ params })
-      render(jsx)
-
-      const scriptTag = document.querySelector('script[type="application/ld+json"]')
-      const jsonLd = JSON.parse(scriptTag?.textContent ?? '{}')
-
-      expect(jsonLd.image).toBe('/og-image.png')
-    })
-
-    it('includes correct URL structure', async () => {
-      const mockPost = {
-        id: '123',
-        slug: 'my-awesome-post',
-        title: 'My Awesome Post',
-        description: 'Test description',
-        firstPublished: '2024-01-15',
-        lastEditedTime: '2024-01-15T00:00:00.000Z',
-        featuredImage: null,
-        blocks: [],
-        prevPost: null,
-        nextPost: null,
-      }
-
-      vi.mocked(getPost).mockResolvedValue(Ok(mockPost))
-
-      const params = Promise.resolve({ slug: 'my-awesome-post' })
-      const jsx = await DynamicRoute({ params })
-      render(jsx)
-
-      const scriptTag = document.querySelector('script[type="application/ld+json"]')
-      const jsonLd = JSON.parse(scriptTag?.textContent ?? '{}')
-
-      expect(jsonLd.url).toBe('https://michaeluloth.com/my-awesome-post/')
-    })
-  })
-
   describe('success cases', () => {
     it('fetches and renders post with blocks and navigation', async () => {
       const mockPost = {
