+++ app/[slug]/page.tsx
@@ -3,9 +3,9 @@ import { notFound } from 'next/navigation'
 
 import getPost from '@/io/notion/getPost'
 import getPosts from '@/io/notion/getPosts'
-import { SITE_URL, SITE_NAME, SITE_AUTHOR, TWITTER_HANDLE, DEFAULT_OG_IMAGE } from '@/seo/constants'
+import { SITE_NAME, SITE_AUTHOR, TWITTER_HANDLE, DEFAULT_OG_IMAGE } from '@/seo/constants'
 import { transformCloudinaryForOG } from '@/io/cloudinary/ogImageTransforms'
-import type { Post as PostType } from '@/io/notion/schemas/post'
+import { generateArticleJsonLd, getPostUrl } from '@/seo/article-json-ld'
 
 import PageLayout from '@/ui/layouts/page-layout'
 import Post from '@/ui/post/post'
@@ -27,87 +27,12 @@ export async function generateStaticParams(): Promise<Params[]> {
   return posts.map(post => ({ slug: post.slug }))
 }
 
-type Props = Readonly<{
-  params: Promise<Params>
-}>
-
-export default async function DynamicRoute({ params }: Props) {
-  const slug = (await params).slug
-
-  const post = (await getPost({ slug, includeBlocks: true, includePrevAndNext: true })).unwrap()
-
-  if (!post) {
-    notFound()
-  }
-
-  const jsonLd = generateJsonLd(post, slug)
-
-  return (
-    <PageLayout>
-      <Post post={post} prevPost={post.prevPost} nextPost={post.nextPost} />
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{
-          // Escape < to prevent XSS if content contains </script>
-          __html: JSON.stringify(jsonLd).replace(/</g, '\\u003c'),
-        }}
-      />
-    </PageLayout>
-  )
-}
-
-/**
- * Constructs the full URL for a blog post.
- */
-function getPostUrl(slug: string): string {
-  return `${SITE_URL}${slug}/`
-}
-
-type JsonLdArticle = {
-  '@context': string
-  '@type': string
-  headline: string
-  description: string
-  datePublished: string
-  dateModified: string
-  author: {
-    '@type': string
-    name: string
-  }
-  image: string
-  url: string
-}
-
-/**
- * Generates JSON-LD structured data for a blog post.
- * @see https://schema.org/Article
- */
-function generateJsonLd(post: PostType, slug: string): JsonLdArticle {
-  const url = getPostUrl(slug)
-  const rawImage = post.featuredImage ?? DEFAULT_OG_IMAGE
-  const image = transformCloudinaryForOG(rawImage)
-
-  return {
-    '@context': 'https://schema.org',
-    '@type': 'Article',
-    headline: post.title,
-    description: post.description,
-    datePublished: post.firstPublished,
-    dateModified: post.lastEditedTime,
-    author: {
-      '@type': 'Person',
-      name: SITE_AUTHOR,
-    },
-    image,
-    url,
-  }
-}
-
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
   const slug = (await params).slug
   const post = (await getPost({ slug })).unwrap()
 
   if (!post) {
+    // TODO: confirm if this is the right behaviour; what pages would this apply to?
     return {}
   }
 
@@ -125,7 +50,7 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
       type: 'article',
       url,
       siteName: SITE_NAME,
-      locale: 'en_CA',
+      locale: 'en_CA', // TODO: replace with SITE_LOCALE
       title: post.title,
       description: post.description,
       publishedTime: post.firstPublished,
@@ -141,4 +66,32 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
       images: [ogImage],
     },
   }
+}
+
+type Props = Readonly<{
+  params: Promise<Params>
+}>
+
+export default async function DynamicRoute({ params }: Props) {
+  const slug = (await params).slug
+  const post = (await getPost({ slug, includeBlocks: true, includePrevAndNext: true })).unwrap()
+
+  if (!post) {
+    notFound()
+  }
+
+  const jsonLd = generateArticleJsonLd(post, slug)
+
+  return (
+    <PageLayout>
+      <Post post={post} prevPost={post.prevPost} nextPost={post.nextPost} />
+      <script
+        type="application/ld+json"
+        dangerouslySetInnerHTML={{
+          // Escape < to prevent XSS if content contains </script>
+          __html: JSON.stringify(jsonLd).replace(/</g, '\\u003c'),
+        }}
+      />
+    </PageLayout>
+  )
 }
